authorization_service/requirements.txt
aws-cdk-lib==2.148.1
constructs>=10.0.0,<11.0.0
python-dotenv